requirements.txt
numpy==1.19.1
streamlit==0.62.0
pandas==1.1.0
plotly-express==4.9.0
pydeck==0.4.0
seaborn==0.9.0
folium==0.10.0
urllib3==1.25.10
seaborn==0.10.1
matplotlib==3.3.0